oled/requirements.txt
luma.oled>=3.13.0
luma.core>=2.4.2
Pillow>=10.0.0
